reverser moved into blobstore, but leaked all over the place
